Split the Migration and AutoReverseMigration classes so that expressions that cannot be auto-reversed can be marked obsolete.
