+++ src/FluentMigrator/AutoReversingMigration.cs
@@ -16,12 +16,16 @@
 //
 #endregion
 
+using System;
 using System.Linq;
+using FluentMigrator.Builders.Delete;
+using FluentMigrator.Builders.Execute;
+using FluentMigrator.Builders.Update;
 using FluentMigrator.Infrastructure;
 
 namespace FluentMigrator
 {
-    public abstract class AutoReversingMigration : Migration
+    public abstract class AutoReversingMigration : MigrationBase
     {
         public sealed override void Down()
         {
@@ -31,6 +35,24 @@ namespace FluentMigrator
         {
             GetUpExpressions(context);
             context.Expressions = context.Expressions.Select(e => e.Reverse()).Reverse().ToList();
+        }
+
+        [Obsolete("Delete cannot auto-reversed and will no longer be available for auto-reversing migrations.", false)]
+        public IDeleteExpressionRoot Delete
+        {
+            get { return new DeleteExpressionRoot(_context); }
+        }
+
+        [Obsolete("Execute cannot auto-reversed and will no longer be available for auto-reversing migrations.", false)]
+        public IExecuteExpressionRoot Execute
+        {
+            get { return new ExecuteExpressionRoot(_context); }
+        }
+
+        [Obsolete("Update cannot auto-reversed and will no longer be available for auto-reversing migrations.", false)]
+        public IUpdateExpressionRoot Update
+        {
+            get { return new UpdateExpressionRoot(_context); }
         }
     }
 }
