+++ src/FluentMigrator/Migration.cs
@@ -16,76 +16,17 @@
 //
 #endregion
 
-using FluentMigrator.Builders.Alter;
-using FluentMigrator.Builders.Create;
 using FluentMigrator.Builders.Delete;
-using FluentMigrator.Builders.IfDatabase;
-using FluentMigrator.Builders.Insert;
-using FluentMigrator.Builders.Rename;
-using FluentMigrator.Builders.Schema;
-using FluentMigrator.Infrastructure;
 using FluentMigrator.Builders.Execute;
 using FluentMigrator.Builders.Update;
 
 namespace FluentMigrator
 {
-    public abstract class Migration : IMigration
+    public abstract class Migration : MigrationBase
     {
-        private IMigrationContext _context;
-        private readonly object _mutex = new object();
-
-        public abstract void Up();
-        public abstract void Down();
-
-        public void ApplyConventions(IMigrationContext context)
-        {
-            foreach (var expression in context.Expressions)
-                expression.ApplyConventions(context.Conventions);
-        }
-
-        public virtual void GetUpExpressions(IMigrationContext context)
-        {
-            lock (_mutex)
-            {
-                _context = context;
-                Up();
-                _context = null;
-            }
-        }
-
-        public virtual void GetDownExpressions(IMigrationContext context)
-        {
-            lock (_mutex)
-            {
-                _context = context;
-                Down();
-                _context = null;
-            }
-        }
-
-        public IAlterExpressionRoot Alter
-        {
-            get { return new AlterExpressionRoot(_context); }
-        }
-
-        public ICreateExpressionRoot Create
-        {
-            get { return new CreateExpressionRoot(_context); }
-        }
-
         public IDeleteExpressionRoot Delete
         {
             get { return new DeleteExpressionRoot(_context); }
-        }
-
-        public IRenameExpressionRoot Rename
-        {
-            get { return new RenameExpressionRoot(_context); }
-        }
-
-        public IInsertExpressionRoot Insert
-        {
-            get { return new InsertExpressionRoot(_context); }
         }
 
         public IExecuteExpressionRoot Execute
@@ -93,19 +34,9 @@ namespace FluentMigrator
             get { return new ExecuteExpressionRoot(_context); }
         }
 
-        public ISchemaExpressionRoot Schema
-        {
-            get { return new SchemaExpressionRoot(_context); }
-        }
-
         public IUpdateExpressionRoot Update
         {
             get { return new UpdateExpressionRoot(_context); }
-        }
-
-        public IIfDatabaseExpressionRoot IfDatabase(params string[] databaseType)
-        {
-            return new IfDatabaseExpressionRoot(_context, databaseType);
         }
     }
 }
